Road scene with car
Ok so the issue with why we were having so many compiling issues was because of that specific asset of the city. I found a new one so I think I fixed it but it completely messed up the fort from rhino 3d file. The new project is within that folder and is called "stillfixing_safari".